+++ pages/chat-list/all.tsx
@@ -1,23 +1,17 @@
 import { useEffect, useState } from 'react';
-import { useRouter } from 'next/router';
 import { Chat } from '@/@types/types';
-import Image from 'next/image';
-import { useRecoilValue } from 'recoil';
-import { userIdState } from '@/recoil/atoms/userIdState';
-import formatTime from '@/utils/timeFormat';
 import { sortChatList } from '@/utils/chatList';
 import useConnectServerSocket from '@/hooks/useConnectServerSocket';
 import Header from '@/components/Header/Header';
 import {
   ChatListModal,
   CreateChatButton,
-  JudgeWrapper,
+  AllChatListItem,
 } from '@/components/ChatList';
 import chatListAPI from '../../apis/chatListAPI';
 import styles from './ChatList.module.scss';
 
 export default function AllChatList() {
-  const router = useRouter();
   const [isModal, setIsModal] = useState(false);
   const [allChatList, setAllChatList] = useState<Chat[]>([]);
   const getAllChat = async () => {
@@ -33,24 +27,6 @@ export default function AllChatList() {
 
     return () => clearInterval(timer);
   }, []);
-
-  const participateChat = async (e: React.MouseEvent<HTMLButtonElement>) => {
-    if (e.target instanceof HTMLButtonElement) {
-      await chatListAPI.participateChat(e.target.id);
-      router.push({
-        pathname: `/chat/${e.target.id}`,
-        query: { name: e.target.name },
-      });
-    }
-  };
-
-  const userId = useRecoilValue(userIdState);
-  const checkIncluded = (element: { id: string }) => {
-    if (element.id === userId) {
-      return true;
-    }
-    return false;
-  };
 
   const handleModal = () => {
     setIsModal(!isModal);
@@ -71,56 +47,9 @@ export default function AllChatList() {
       <CreateChatButton setIsModal={setIsModal} />
       {isModal && <ChatListModal handleModal={handleModal} />}
       <ul>
-        {allChatList.map(chat => {
-          const { timeDiffText, className } = formatTime(chat.updatedAt);
-          const isincluded = chat.users.some(checkIncluded);
-          return (
-            <li key={chat.id}>
-              <JudgeWrapper isincluded={isincluded} chatId={chat.id}>
-                <Image
-                  alt={`${chat.users[0].username}의 프로필 사진`}
-                  src={chat.users[0].picture}
-                  width={45}
-                  height={45}
-                  className={styles.user_profile}
-                />
-                <div className={styles.chatInfo}>
-                  <div className={styles.chatWrap}>
-                    <div className={styles.chatNameWrap}>
-                      <div className={styles.chatName}>{chat.name}</div>
-                      <span className={styles['user-length']}>
-                        {chat.users.length}
-                      </span>
-                      <div className={styles.chat_updated}>
-                        <span className={styles[className]}>
-                          {timeDiffText}
-                        </span>
-                      </div>
-                    </div>
-                    <div className={styles.chatLastestMesaage}>
-                      {chat.latestMessage?.text}
-                    </div>
-                  </div>
-                  <div className={styles.right}>
-                    {/* <div className={styles.chat_updated}>
-                    {isToday === dateString ? formattedTime : `${dateString}`}
-                  </div> */}
-                    {!isincluded && (
-                      <button
-                        type="button"
-                        id={chat.id}
-                        name={chat.name}
-                        onClick={participateChat}
-                      >
-                        참여
-                      </button>
-                    )}
-                  </div>
-                </div>
-              </JudgeWrapper>
-            </li>
-          );
-        })}
+        {allChatList.map(chat => (
+          <AllChatListItem key={chat.id} chat={chat} />
+        ))}
       </ul>
     </div>
   );
